fastapi_auth/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==3.2.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kafka==1.3.5
kafka-python==2.0.2
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.18.1
PyJWT==2.8.0
pymongo==4.6.3
pytest==8.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing-extensions==4.11.0
uvicorn==0.29.0
